♻️(minio) dissociate console and api ingress
We want to dissociate the api and console ingress. Since minio decided
to remove almost everything fron the admin console, having it as main
ingress does not make sens. So the main ingress is dedicated to the API
and can be configured using the ingress object, by default it is
disabled. The console ingress can be configured using the consoleIngress
object and it is disabled by default.